+++ src/main/java/Idea/Archive/IdeaArchive/global/security/SecurityConfig.java
@@ -47,6 +47,7 @@ public class SecurityConfig {
                 .antMatchers("/member/**").authenticated()
                 .antMatchers("/post/**").authenticated()
                 .antMatchers("/post/comment/**").authenticated()
+                .antMatchers("/google/**").authenticated()
 
                 .antMatchers("/member/notice/**").hasAuthority("ADMIN")
                 .anyRequest().authenticated();
